Local communication implement
This change implemented the use of local database(SQLite) and
communication between users (one to one) using MultipeerConnectivity
class.

+++ JAM/JAM/EBCardView/EBCardCollectionViewLayout.m
@@ -37,7 +37,7 @@ static NSString * const CellKind = @"CardCell";
 }
 
 - (NSString *)keyForIndexPath:(NSIndexPath *)indexPath {
-    NSString *retVal = [NSString stringWithFormat:@"%d-%d", indexPath.section, indexPath.row];
+    NSString *retVal = [NSString stringWithFormat:@"%ld-%ld", (long)indexPath.section, (long)indexPath.row];
     return retVal;
 }
